+++ src/Extractors/ColumnPicker.php
@@ -8,9 +8,10 @@ use Generator;
 /**
  * Column Picker
  *
- * An extractor that wraps another iterator and selects and rearranges the columns of the data
- * table according to the user-specified keys. The key of a column may either be a string or a
- * column number (integer) depending on the base iterator.
+ * An extractor that wraps another iterator and selects and reorders the columns of the data
+ * table according to the keys specified by the user. The key of a column may either be a
+ * string or a column number (integer) depending on the way the columns are indexed in the
+ * base iterator.
  *
  * @category    Machine Learning
  * @package     Rubix/ML